move util modules to bofire.utils
