+++ bofire/data_models/domain/api.py
@@ -1,4 +1,3 @@
 from bofire.data_models.domain.constraints import Constraints  # noqa: F401
 from bofire.data_models.domain.domain import Domain  # noqa: F401
 from bofire.data_models.domain.features import Features, Inputs, Outputs  # noqa: F401
-from bofire.data_models.domain.subdomain import get_subdomain  # noqa: F401
