+++ bofire/surrogates/single_task_gp.py
@@ -12,8 +12,8 @@ from bofire.data_models.enum import CategoricalEncodingEnum, OutputFilteringEnum
 from bofire.data_models.surrogates.api import SingleTaskGPSurrogate as DataModel
 from bofire.data_models.surrogates.scaler import ScalerEnum
 from bofire.surrogates.botorch import BotorchSurrogate
-from bofire.surrogates.torch_tools import tkwargs
 from bofire.surrogates.trainable import TrainableSurrogate
+from bofire.utils.torch_tools import tkwargs
 
 
 def get_dim_subsets(d: int, active_dims: List[int], cat_dims: List[int]):
